requirements.txt
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
yfinance>=0.2.0
ccxt>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
python-dotenv>=1.0.0
requests>=2.31.0
ta>=0.10.0
backtesting>=0.3.3
jupyter>=1.0.0
streamlit>=1.28.0
kiteconnect>=4.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
mysql-connector-python>=8.1.0
scipy>=1.10.0
joblib>=1.3.0
